APPS-1232 - review changes, moved parse_obj from cli.exec_io to utils.resolver.
