+++ src/python/dxpy/utils/resolver.py
@@ -29,9 +29,9 @@ import os, sys, json, re
 
 import dxpy
 from .describe import get_ls_l_desc
-from ..exceptions import DXError
 from ..compat import str, input, basestring
 from ..cli import try_call, INTERACTIVE_CLI
+from ..exceptions import DXCLIError, DXError
 
 def pick(choices, default=None, str_choices=None, prompt=None, allow_mult=False, more_choices=False):
     '''
@@ -1367,3 +1367,17 @@ def parse_input_keyval(keyeqval):
         return (name, value)
     except:
         raise DXCLIError('An input was found that did not conform to the syntax: -i<input name>=<input value>')
+
+def parse_obj(string, klass):
+    if string == '':
+        raise DXCLIError('Error: Nonempty string cannot be resolved')
+    project, path, entity_result = resolve_existing_path(string)
+    if entity_result is None:
+        raise DXCLIError('Could not resolve \"' + string + '\" to a name or ID')
+    if not entity_result['describe']['class'] == klass:
+        raise DXCLIError('Error: The given object is of class ' + entity_result['describe']['class'] + ' but an object of class ' + klass + ' was expected.')
+    if is_hashid(string):
+        return {'$dnanexus_link': entity_result['id']}
+    else:
+        return {'$dnanexus_link': {"project": entity_result['describe']['project'],
+                                   "id": entity_result['id']}}
